feat!: remove cs_comments_service support for forum's subscription APIs - This will force the use of the new v2 forum's APIs for subscriptions.
