+++ lms/djangoapps/discussion/rest_api/tests/test_api_v2.py
@@ -282,6 +282,18 @@ class CreateThreadTest(
         }
         self.check_mock_called_with("update_thread_flag", -1, **params)
 
+    def test_following(self):
+        self.register_post_thread_response({"id": "test_id", "username": self.user.username})
+        self.register_subscription_response(self.user)
+        data = self.minimal_data.copy()
+        data["following"] = "True"
+        result = create_thread(self.request, data)
+        assert result['following'] is True
+        self.check_mock_called("create_subscription")
+
+        params = {'user_id': str(self.user.id), 'course_id': str(self.course.id), 'source_id': 'test_id'}
+        self.check_mock_called_with("create_subscription", 0, **params)
+
 
 @ddt.ddt
 @disable_signal(api, "comment_created")
@@ -711,6 +723,49 @@ class UpdateThreadTest(
                 vote_count -= 1
                 assert result["vote_count"] == vote_count
                 self.register_get_user_response(self.user, upvoted_ids=[])
+
+    @ddt.data(*itertools.product([True, False], [True, False]))
+    @ddt.unpack
+    @mock.patch("eventtracking.tracker.emit")
+    def test_following(self, old_following, new_following, mock_emit):
+        """
+        Test attempts to edit the "following" field.
+
+        old_following indicates whether the thread should be followed at the
+        start of the test. new_following indicates the value for the "following"
+        field in the update. If old_following and new_following are the same, no
+        update should be made. Otherwise, a subscription should be POSTed or
+        DELETEd according to the new_following value.
+        """
+        if old_following:
+            self.register_get_user_response(self.user, subscribed_thread_ids=["test_thread"])
+        self.register_subscription_response(self.user)
+        self.register_thread()
+        data = {"following": new_following}
+        signal_name = "thread_followed" if new_following else "thread_unfollowed"
+        mock_path = f"openedx.core.djangoapps.django_comment_common.signals.{signal_name}.send"
+        with mock.patch(mock_path) as signal_patch:
+            result = update_thread(self.request, "test_thread", data)
+            if old_following != new_following:
+                self.assertEqual(signal_patch.call_count, 1)
+        assert result['following'] == new_following
+
+        if old_following == new_following:
+            assert not self.check_mock_called("create_subscription")
+        else:
+            params = {'user_id': str(self.user.id), 'course_id': str(self.course.id), 'source_id': 'test_thread'}
+            if new_following:
+                assert self.check_mock_called("create_subscription")
+            else:
+                assert self.check_mock_called("delete_subscription")
+
+            event_name, event_data = mock_emit.call_args[0]
+            expected_event_action = 'followed' if new_following else 'unfollowed'
+            assert event_name == f'edx.forum.thread.{expected_event_action}'
+            assert event_data['commentable_id'] == 'original_topic'
+            assert event_data['id'] == 'test_thread'
+            assert event_data['followed'] == new_following
+            assert event_data['user_forums_roles'] == ['Student']
 
 
 @ddt.ddt
