+++ openedx/core/djangoapps/django_comment_common/comment_client/user.py
@@ -52,39 +52,19 @@ class User(models.Model):
 
     def follow(self, source, course_id=None):
         course_key = utils.get_course_key(self.attributes.get("course_id") or course_id)
-        if is_forum_v2_enabled(course_key):
-            forum_api.create_subscription(
-                user_id=self.id,
-                source_id=source.id,
-                course_id=str(course_key)
-            )
-        else:
-            params = {'source_type': source.type, 'source_id': source.id}
-            utils.perform_request(
-                'post',
-                _url_for_subscription(self.id),
-                params,
-                metric_action='user.follow',
-                metric_tags=self._metric_tags + [f'target.type:{source.type}'],
-            )
+        forum_api.create_subscription(
+            user_id=self.id,
+            source_id=source.id,
+            course_id=str(course_key)
+        )
 
     def unfollow(self, source, course_id=None):
         course_key = utils.get_course_key(self.attributes.get("course_id") or course_id)
-        if is_forum_v2_enabled(course_key):
-            forum_api.delete_subscription(
-                user_id=self.id,
-                source_id=source.id,
-                course_id=str(course_key)
-            )
-        else:
-            params = {'source_type': source.type, 'source_id': source.id}
-            utils.perform_request(
-                'delete',
-                _url_for_subscription(self.id),
-                params,
-                metric_action='user.unfollow',
-                metric_tags=self._metric_tags + [f'target.type:{source.type}'],
-            )
+        forum_api.delete_subscription(
+            user_id=self.id,
+            source_id=source.id,
+            course_id=str(course_key)
+        )
 
     def vote(self, voteable, value, course_id=None):
         course_key = utils.get_course_key(self.attributes.get("course_id") or course_id)
